+++ assets/js/modules/utils.js
@@ -1,0 +1,10 @@
+/*
+ * General utility functions
+ */
+
+
+function array_contains(arr, str) {
+    return arr.indexOf(str) > -1;
+}
+
+export { array_contains };
